Test: 댓글 도메인 테스트 분리, 공통 fake 엔티티 추출
