+++ src/test/java/com/gdsc_knu/official_homepage/comment/service/CommentServiceTest.java
@@ -1,4 +1,4 @@
-package com.gdsc_knu.official_homepage.post.comment;
+package com.gdsc_knu.official_homepage.comment.service;
 
 import com.gdsc_knu.official_homepage.dto.PagingResponse;
 import com.gdsc_knu.official_homepage.dto.post.CommentRequest;
@@ -24,6 +24,7 @@ import org.springframework.data.domain.PageRequest;
 import java.util.Collections;
 import java.util.Optional;
 
+import static com.gdsc_knu.official_homepage.comment.CommentTestEntityFactory.*;
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.*;
 import static org.assertj.core.api.Assertions.*;
@@ -39,7 +40,7 @@ class CommentServiceTest {
     private final Long memberId = 1L;
     private final Long postId = 1L;
     private final Optional<Member> author = createAuthor(memberId);
-    private final Optional<Post> post = createPost(memberId);
+    private final Optional<Post> post = createPost(postId, memberId);
 
     @Test
     @DisplayName("댓글이 정상적으로 저장된다")
@@ -54,21 +55,12 @@ class CommentServiceTest {
     }
 
 
-    @Test
-    @DisplayName("댓글이 저장될때 게시글의 댓글 개수 필드가 정상적으로 업데이트되고, 게시글 댓글 컬렉션에 포함된다.")
-    void saveCommentCount() {
-        // given
-        Comment comment1 = Comment.from("댓글 내용", author.orElseThrow(), post.orElseThrow(), null);
-        // when && then
-        assertThat(post.get().getCommentCount()).isEqualTo(1);
-        assertThat(post.get().getCommentList()).containsExactly(comment1);
-    }
 
     @Test
     @DisplayName("존재하지 않는 댓글에 답글을 남기면 오류를 발생시킨다.")
     void saveCommentInvalidParent() {
         // given
-        when(postRepository.findById(postId)).thenReturn(createPost(memberId));
+        when(postRepository.findById(postId)).thenReturn(createPost(postId, memberId));
         when(memberRepository.findById(memberId)).thenReturn(createAuthor(memberId));
         Long notExistCommentId = 1L;
         CommentRequest.Create request = new CommentRequest.Create(notExistCommentId, "댓글 내용");
@@ -86,8 +78,8 @@ class CommentServiceTest {
     void getCommentByCommentAuthor() {
         // given (댓글을 본인이 작성했으나, 게시글은 본인이 작성한게 아닌 경우)
         long postAuthorId = 2L;
-        when(postRepository.findById(postId)).thenReturn(createPost(postAuthorId));
-        Comment comment = createComment(memberId);
+        when(postRepository.findById(postId)).thenReturn(createPost(postId, postAuthorId));
+        Comment comment = createComment(postId, memberId);
         when(commentRepository.findCommentAndReply(PageRequest.of(0,5), postId))
                 .thenReturn(new PageImpl<>(Collections.singletonList(comment)));
         // when
@@ -104,8 +96,8 @@ class CommentServiceTest {
     void getCommentByPostAuthor() {
         //given (본인의 게시글에 다른 사람의 댓글이 존재하는 경우)
         long commentAuthorId = 3L;
-        when(postRepository.findById(postId)).thenReturn(createPost(memberId));
-        Comment comment = createComment(commentAuthorId);
+        when(postRepository.findById(postId)).thenReturn(createPost(postId, memberId));
+        Comment comment = createComment(postId, commentAuthorId);
         when(commentRepository.findCommentAndReply(PageRequest.of(0,5), postId))
                 .thenReturn(new PageImpl<>(Collections.singletonList(comment)));
         // when
@@ -118,19 +110,6 @@ class CommentServiceTest {
     }
 
 
-    @Test
-    @DisplayName("상위 댓글 삭제 시 연관된 하위 댓글 삭제까지의 댓글개수가 정상적으로 업데이트된다.")
-    void deleteParentComment() {
-        // given
-        Comment parent = Comment.from("댓글 내용", author.orElseThrow(), post.orElseThrow(), null);
-        Comment child1 = Comment.from("댓글 내용", author.orElseThrow(), post.orElseThrow(), parent);
-        Comment child2 = Comment.from("댓글 내용", author.orElseThrow(), post.orElseThrow(), parent);
-        // when
-        parent.delete();
-
-        //then
-        assertThat(post.get().getCommentCount()).isEqualTo(0);
-    }
 
 
 
@@ -138,33 +117,6 @@ class CommentServiceTest {
 
 
 
-    private Optional<Post> createPost(long authorId) {
-        return Optional.ofNullable(Post.builder()
-                .id(postId)
-                .member(createAuthor(authorId).get())
-                .build());
-    }
 
-    private Comment createComment(long authorId) {
-        Member author = createAuthor(authorId).get();
-        Comment parent = Comment.builder()
-                .id(1L)
-                .build();
-        return Comment.builder()
-                .id(postId)
-                .content("댓글")
-                .author(author)
-                .parent(parent)
-                .build();
-    }
 
-    private Optional<Member> createAuthor(long id) {
-        return Optional.ofNullable(Member.builder()
-                .id(id)
-                .email("email@email.com")
-                .phoneNumber("010-0000-0000")
-                .name("테스트 유저")
-                .profileUrl("image.png")
-                .build());
-    }
 }
